fix: UseCase の package の配置箇所とファイル内の package 名が一致していないため、when 式で網羅性を担保できない問題を引き起こしていたため、一致するように rename した
