+++ src/main/kotlin/com/example/realworldkotlinspringbootjdbc/presentation/UserAndAuthenticationController.kt
@@ -7,9 +7,9 @@ import com.example.realworldkotlinspringbootjdbc.presentation.response.CurrentUs
 import com.example.realworldkotlinspringbootjdbc.presentation.response.serializeMyErrorListForResponseBody
 import com.example.realworldkotlinspringbootjdbc.presentation.response.serializeUnexpectedErrorForResponseBody
 import com.example.realworldkotlinspringbootjdbc.presentation.shared.AuthorizationError
-import com.example.realworldkotlinspringbootjdbc.usecase.LoginUseCase
-import com.example.realworldkotlinspringbootjdbc.usecase.RegisterUserUseCase
-import com.example.realworldkotlinspringbootjdbc.usecase.UpdateUserUseCase
+import com.example.realworldkotlinspringbootjdbc.usecase.user_and_authentication.LoginUseCase
+import com.example.realworldkotlinspringbootjdbc.usecase.user_and_authentication.RegisterUserUseCase
+import com.example.realworldkotlinspringbootjdbc.usecase.user_and_authentication.UpdateUserUseCase
 import com.example.realworldkotlinspringbootjdbc.util.MyAuth
 import com.example.realworldkotlinspringbootjdbc.util.MySession
 import com.example.realworldkotlinspringbootjdbc.util.MySessionJwt
