+++ src/main/kotlin/com/example/realworldkotlinspringbootjdbc/usecase/comment/ListCommentUseCase.kt
@@ -1,4 +1,4 @@
-package com.example.realworldkotlinspringbootjdbc.usecase
+package com.example.realworldkotlinspringbootjdbc.usecase.comment
 
 import arrow.core.Either
 import arrow.core.Either.Left
